fix tampilan

+++ app/Http/Controllers/CategoryController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Category;
+use App\Models\Product;
 use Illuminate\Support\Facades\Validator;
 use Yajra\DataTables\DataTables;
 
@@ -52,13 +53,15 @@ class CategoryController extends Controller
         }
     }
 
-    public function edit($id){
-      $data = Category::query()
-        ->select ('id','nama')->where('id',$id)->first();
+    public function edit($id)
+    {
+        $data = Category::query()
+            ->select('id', 'nama')->where('id', $id)->first();
         return response()->json(['data' => $data]);
     }
 
-    public function update(Request $request){
+    public function update(Request $request)
+    {
         $validator = Validator::make($request->all(), [
             'category' => ['required']
         ], ['category.required' => 'Mohon Inputkan Category']);
@@ -69,23 +72,29 @@ class CategoryController extends Controller
 
         //simpan data ke db
         $data = Category::query()
-        ->where('id',$request->id)
-        ->update([
-            'nama' => $request->category
-        ]);
+            ->where('id', $request->id)
+            ->update([
+                'nama' => $request->category
+            ]);
         if ($data) {
             return response()->json(['success' => 1]);
         } else {
             return response()->json(['success' => 0]);
-        } 
+        }
     }
 
     // modal delete
-    public function delete($id){
-        Category::where("id", $id)->delete();
-        $response['succes']= true;
-        $response['message']= "Data berhasil dihapus";
+    public function delete($id)
+    {
+        $findCateorieProduct = Product::where('id_category', $id)->first();
+        if ($findCateorieProduct) {
+            $response['succes'] = false;
+            $response['message'] = "Categori sedang digunakan dalam produk";
+        } else {
+            Category::where("id", $id)->delete();
+            $response['succes'] = true;
+            $response['message'] = "Data berhasil dihapus";
+        }
         return response()->json(['data' => $response]);
     }
-
 }
